Finished the GUI

+++ src/org/warp/midito3d/midi/Note.java
@@ -1,11 +1,11 @@
 package org.warp.midito3d.midi;
 
 public class Note {
-	public final int note;
-	public final int velocity;
+	public final double note;
+	public final double velocity;
 	public final long startTick;
 	
-	public Note(int note, int velocity, long startTick) {
+	public Note(double note, double velocity, long startTick) {
 		this.note = note;
 		this.velocity = velocity;
 		this.startTick = startTick;
